requirements.txt
sqlalchemy~=2.0.38
environs~=14.1.1
asyncpg~=0.30.0
uvicorn~=0.34.0
fastapi~=0.115.8
python-jose~=3.4.0
pydantic~=2.10.6